Improve performances of project/packages retrieval and Remove-ItemSilently.

+++ .build/scripts/Pask.ps1
@@ -236,12 +236,7 @@ function script:Remove-ItemSilently {
         if ($ItemObject -is [System.IO.FileInfo]) {
             Remove-Item -Path $ItemObject.FullName -Force | Out-Null
         } elseif ($ItemObject -is [System.IO.DirectoryInfo]) {
-            # Ensure removal of directories exceeding the 260 characters limit
-            Get-ChildItem -Path $ItemObject -Recurse `
-                | Sort -Descending @{Expression = {$_.FullName.Length}} `
-                | Select -ExpandProperty FullName `
-                | Remove-ItemSilently
-            CMD /C ("RD /S /Q ""{0}""" -f $ItemObject.FullName)
+            [System.IO.Directory]::Delete(("{0}" -f $ItemObject.FullName), $true)
         } else {
             $ItemObject.FullName | Remove-ItemSilently
         }
@@ -287,9 +282,7 @@ function script:Initialize-NuGetExe {
    The full path
 #> 
 function script:Get-PackagesDir {
-    if(Pask-Cache $MyInvocation.MyCommand.Name) {
-        return Pask-Cache $MyInvocation.MyCommand.Name
-    }
+    if(Pask-Cache $MyInvocation.MyCommand.Name) { return Pask-Cache $MyInvocation.MyCommand.Name }
     $PackagesDir = Join-Path $PaskFullPath "packages"
     $NuGet = Get-NuGetExe
     if (Test-Path $NuGet) {
@@ -335,10 +328,12 @@ function script:Restore-NuGetDevelopmentPackages {
     $PackagesDir = Get-PackagesDir
     $NuGetExe = Get-NuGetExe
     Get-SolutionPackages | Where { $_.developmentDependency -eq $true } | ForEach {
-        Invoke-Command -ErrorAction Stop -ScriptBlock { & $NuGetExe install $_.id -Version $_.version -OutputDirectory "$PackagesDir" -NonInteractive -Verbosity quiet }
-        if ($LastExitCode) {
-            Write-BuildMessage -Message "Error restoring NuGet package $($_.id).$($_.version)" -ForegroundColor "Red"
-            exit $LastExitCode
+        if(-not (Test-Path (Get-PackageDir $_.id))) {
+            Invoke-Command -ErrorAction Stop -ScriptBlock { & $NuGetExe install $_.id -Version $_.version -OutputDirectory "$PackagesDir" -NonInteractive -Verbosity quiet }
+            if ($LastExitCode) {
+                Write-BuildMessage -Message "Error restoring NuGet package $($_.id).$($_.version)" -ForegroundColor "Red"
+                exit $LastExitCode
+            }
         }
     }
 }
@@ -437,6 +432,7 @@ function script:Get-ProjectFullName {
    )
 #>
 function script:Get-SolutionProjects {
+    if(Pask-Cache $MyInvocation.MyCommand.Name) { return Pask-Cache $MyInvocation.MyCommand.Name }
     $Projects = @()
     Get-Content "$SolutionFullName" |
     Select-String 'Project\(' |
@@ -451,6 +447,7 @@ function script:Get-SolutionProjects {
                 }
             }
         }
+    Pask-Cache $MyInvocation.MyCommand.Name -Value $Projects
     return $Projects
 }
 
@@ -468,6 +465,8 @@ function script:Get-SolutionProjects {
    )
 #>
 function script:Get-SolutionPackages {
+    if(Pask-Cache $MyInvocation.MyCommand.Name) { return Pask-Cache $MyInvocation.MyCommand.Name }
+    
     $Packages = @()
 
     foreach($Project in Get-SolutionProjects) {
@@ -477,7 +476,9 @@ function script:Get-SolutionPackages {
         }
     }
 
-    return ($Packages | Select -Unique id, version, developmentDependency | Sort id, version)
+    $SolutionPackages = $Packages | Select -Unique id, version, developmentDependency | Sort id, version
+    Pask-Cache $MyInvocation.MyCommand.Name -Value $SolutionPackages
+    return $SolutionPackages
 }
 
 <#
